Fixup framing bugs
There were two bugs...

1. Movement was double stepping. This was due to there being two events fired
for every keypress. (I needed to filter out the key lift event, and only care
about key trigger.)

2. Sometimes when resizing the window, the frame would break in weird ways and
never return to normal. This was a result of a couple of different issues, they
main one being that through neither movement nor resize events did I ever hard
recenter the frame (I only ever incremented it from a known initial position.
In the event of the race condition where multiple resize events come in, but
the one which triggers a frame update is inaccurate, it's possible that the
frame size becomes desynced with the actual center.